update: Read-side DB に口座残高も保存するように変更

+++ app/adapter/src/main/scala/myapp/adapter/account/TransactionDto.scala
@@ -3,7 +3,7 @@ package myapp.adapter.account
 final case class TransactionDto(
     transactionId: String,
     transactionType: String,
-    amount: Long,
-    balance: Long,
+    amount: Int,
+    balance: Int,
     transactedAt: Long,
 )
